clean code, move connection to separate class
