+++ src/main/java/com/example/jakartaeelabs/Database/ManageGroups.java
@@ -1,29 +1,15 @@
+package com.example.jakartaeelabs.Database;
+
 import java.sql.*;
 
 public class ManageGroups {
 
-    /**
-     * Connect to the test.db database
-     *
-     * @return the Connection object
-     */
-    private Connection connect() {
-        // SQLite connection string
-        String url = "jdbc:sqlite:src\\main\\java\\com\\example\\jakartaeelabs\\database\\new.db";
-        Connection conn = null;
-        try {
-            conn = DriverManager.getConnection(url);
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-        return conn;
-    }
+    private final Connection connection = Connect.connect();
 
     public void insert(String name) {
         String sql = "INSERT INTO groups(name) VALUES(?)";
 
-        try (Connection conn = this.connect();
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
+        try (PreparedStatement pstmt = this.connection.prepareStatement(sql)) {
             pstmt.setString(1, name);
             pstmt.executeUpdate();
         } catch (SQLException e) {
@@ -35,9 +21,7 @@ public class ManageGroups {
         String sql = "UPDATE groups SET name = ? "
                 + "WHERE id = ?";
 
-        try (Connection conn = this.connect();
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
-
+        try (PreparedStatement pstmt = this.connection.prepareStatement(sql)) {
             // set the corresponding param
             pstmt.setString(1, name);
             pstmt.setInt(2, id);
@@ -51,9 +35,7 @@ public class ManageGroups {
     public void delete(int id) {
         String sql = "DELETE FROM groups WHERE id = ?";
 
-        try (Connection conn = this.connect();
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
-
+        try (PreparedStatement pstmt = this.connection.prepareStatement(sql)) {
             // set the corresponding param
             pstmt.setInt(1, id);
             // execute the delete statement
@@ -64,16 +46,15 @@ public class ManageGroups {
         }
     }
 
-    public void selectAll(){
+    public void selectAll() {
         String sql = "SELECT id, name FROM groups";
 
-        try (Connection conn = this.connect();
-             Statement stmt  = conn.createStatement();
-             ResultSet rs    = stmt.executeQuery(sql)){
+        try (Statement stmt = this.connection.createStatement();
+             ResultSet rs = stmt.executeQuery(sql)) {
 
             // loop through the result set
             while (rs.next()) {
-                System.out.println(rs.getInt("id") +  "\t" +
+                System.out.println(rs.getInt("id") + "\t" +
                         rs.getString("name"));
             }
         } catch (SQLException e) {
@@ -81,21 +62,20 @@ public class ManageGroups {
         }
     }
 
-    public void getGroup(int id){
+    public void getGroup(int id) {
         String sql = "SELECT id, name "
                 + "FROM groups WHERE id = ?";
 
-        try (Connection conn = this.connect();
-             PreparedStatement pstmt  = conn.prepareStatement(sql)){
+        try (PreparedStatement pstmt = this.connection.prepareStatement(sql)) {
 
             // set the value
-            pstmt.setInt(1,id);
+            pstmt.setInt(1, id);
             //
-            ResultSet rs  = pstmt.executeQuery();
+            ResultSet rs = pstmt.executeQuery();
 
             // loop through the result set
             while (rs.next()) {
-                System.out.println(rs.getInt("id") +  "\t" +
+                System.out.println(rs.getInt("id") + "\t" +
                         rs.getString("name"));
             }
         } catch (SQLException e) {
@@ -103,11 +83,7 @@ public class ManageGroups {
         }
     }
 
-    /**
-     * @param args the command line arguments
-     */
     public static void main(String[] args) {
-
         ManageGroups app = new ManageGroups();
         // insert three new rows
 //        app.insert("IO-21mn");
